🎨 :: rename evening_study to night_study
