+++ domain/src/main/java/kr/hs/dgsw/smartschool/domain/model/night_study/NightStudyStudent.kt
@@ -1,4 +1,4 @@
-package kr.hs.dgsw.smartschool.domain.model.evening_study
+package kr.hs.dgsw.smartschool.domain.model.night_study
 
 data class NightStudyStudent(
     val name: String,
