+++ presentation/src/main/java/kr/hs/dgsw/smartschool/dodamdodam_teacher/features/night_study/mvi/NightStudyContract.kt
@@ -1,7 +1,7 @@
 package kr.hs.dgsw.smartschool.dodamdodam_teacher.features.night_study.mvi
 
 import kr.hs.dgsw.smartschool.domain.model.classroom.Classroom
-import kr.hs.dgsw.smartschool.domain.model.evening_study.NightStudy
+import kr.hs.dgsw.smartschool.domain.model.night_study.NightStudy
 import kr.hs.dgsw.smartschool.domain.model.member.Member
 import kr.hs.dgsw.smartschool.domain.model.member.student.Student
